+++ simple_django_starter/CHANGELOG.md
@@ -1,3 +1,10 @@
 # 0.1.0+1
 
 - Initial release.
+
+# 0.1.0+2
+
+- Added SMTP email support (by default gmail).
+- Added environment variables.
+- Django Admin Theme added.
+- Small Bug fixes.